finish 10 first level data

+++ Assets/Game.Main/manhnd_sdk/Scripts/ExtensionMethods/MeshRendererExtensions.cs
@@ -14,7 +14,7 @@ namespace manhnd_sdk.Scripts.ExtensionMethods
         
         // Config
         private const float OffsetUnit = 0.25f;
-        private const float PiecesPerRow = 4f;
+        private const int PiecesPerRow = 4;
         
         public static void SetVertexLitColor(this MeshRenderer meshRenderer, Color newColor)
         {
@@ -28,7 +28,7 @@ namespace manhnd_sdk.Scripts.ExtensionMethods
             meshRenderer.GetPropertyBlock(mpb);
             
             float offsetX = (byte)colorType % PiecesPerRow * OffsetUnit;
-            float offsetY = -(byte)colorType / PiecesPerRow * OffsetUnit;
+            float offsetY = -(byte)((byte)colorType / PiecesPerRow)* OffsetUnit;
             mpb.SetVector(VertexLitMainTexSt, new Vector4(1, 1, offsetX, offsetY));
             
             meshRenderer.SetPropertyBlock(mpb);
